# docs/openapi/econsent_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cadence Clinical - eConsent",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/econsent/withdrawal": {
      "post": {
        "tags": [
          "Withdrawal"
        ],
        "summary": "Withdraw Subject Consent",
        "description": "Formally revokes/withdraws a subject's consent and records 21 CFR Part 11 audit records.",
        "operationId": "withdraw_subject_consent_api_v1_econsent_withdrawal_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsentWithdrawalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentWithdrawalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/econsent/withdrawal/{study_id}/{subject_pseudonym}": {
      "get": {
        "tags": [
          "Withdrawal"
        ],
        "summary": "Get Subject Withdrawal Status",
        "description": "Retrieves the withdrawal record for a subject if one exists.",
        "operationId": "get_subject_withdrawal_status_api_v1_econsent_withdrawal__study_id___subject_pseudonym__get",
        "parameters": [
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Study Id"
            }
          },
          {
            "name": "subject_pseudonym",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject Pseudonym"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConsentWithdrawalResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/econsent/export/cdisc-odm/{study_id}/{subject_pseudonym}/{template_id}/{version_index}": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Export Subject Cdisc Odm",
        "description": "Exports clinical eConsent records in standard CDISC ODM v1.3.2/v2.0 XML.",
        "operationId": "export_subject_cdisc_odm_api_v1_econsent_export_cdisc_odm__study_id___subject_pseudonym___template_id___version_index__get",
        "parameters": [
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Study Id"
            }
          },
          {
            "name": "subject_pseudonym",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject Pseudonym"
            }
          },
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          },
          {
            "name": "version_index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version Index"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CdiscOdmExportResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/econsent/export/certificate/{study_id}/{subject_pseudonym}/{template_id}/{version_index}": {
      "get": {
        "tags": [
          "Export"
        ],
        "summary": "Export Verifiable Certificate",
        "description": "Generates a standalone, tamper-evident HTML consent certificate.",
        "operationId": "export_verifiable_certificate_api_v1_econsent_export_certificate__study_id___subject_pseudonym___template_id___version_index__get",
        "parameters": [
          {
            "name": "study_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Study Id"
            }
          },
          {
            "name": "subject_pseudonym",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Subject Pseudonym"
            }
          },
          {
            "name": "template_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Template Id"
            }
          },
          {
            "name": "version_index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Version Index"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VerifiableCertificateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CdiscOdmExportResponse": {
        "properties": {
          "study_id": {
            "type": "string",
            "title": "Study Id"
          },
          "subject_pseudonym": {
            "type": "string",
            "title": "Subject Pseudonym"
          },
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "version_index": {
            "type": "integer",
            "title": "Version Index"
          },
          "odm_version": {
            "type": "string",
            "title": "Odm Version"
          },
          "xml_content": {
            "type": "string",
            "title": "Xml Content"
          }
        },
        "type": "object",
        "required": [
          "study_id",
          "subject_pseudonym",
          "template_id",
          "version_index",
          "odm_version",
          "xml_content"
        ],
        "title": "CdiscOdmExportResponse"
      },
...
      "ConsentWithdrawalRequest": {
        "properties": {
          "study_id": {
            "type": "string",
            "title": "Study Id"
          },
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "subject_pseudonym": {
            "type": "string",
            "title": "Subject Pseudonym"
          },
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "withdrawal_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Withdrawal Date"
          },
          "reason_category": {
            "type": "string",
            "title": "Reason Category",
            "description": "Category: Adverse Event, Personal Choice, Relocation, Other"
          },
          "reason_detail": {
            "type": "string",
            "title": "Reason Detail",
            "description": "Specific explanation for withdrawal"
          },
          "scope": {
            "type": "string",
            "title": "Scope",
            "description": "STOP_INTERVENTIONS_ONLY, STOP_ALL_DATA_COLLECTION, DESTROY_SAMPLES",
            "default": "STOP_ALL_DATA_COLLECTION"
          },
          "investigator_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Investigator Id"
          },
          "reason_for_change": {
            "type": "string",
            "title": "Reason For Change",
            "description": "Mandatory change justification"
          }
        },
        "type": "object",
        "required": [
          "study_id",
          "site_id",
          "subject_pseudonym",
          "template_id",
          "reason_category",
          "reason_detail",
          "reason_for_change"
        ],
        "title": "ConsentWithdrawalRequest"
      },
      "ConsentWithdrawalResponse": {
        "properties": {
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "Chronological UTC timestamp when the record was created."
          },
          "created_by": {
            "type": "string",
            "title": "Created By",
            "description": "Unique identifier of the user who created the record."
          },
          "reason_for_change": {
            "type": "string",
            "title": "Reason For Change",
            "description": "Mandatory explanation or audit justification for creating or mutating this record."
          },
          "version_index": {
            "type": "integer",
            "title": "Version Index",
            "description": "Row version counter or index.",
            "default": 1
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "study_id": {
            "type": "string",
            "title": "Study Id"
          },
          "site_id": {
            "type": "string",
            "title": "Site Id"
          },
          "subject_pseudonym": {
            "type": "string",
            "title": "Subject Pseudonym"
          },
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "withdrawal_date": {
            "type": "string",
            "format": "date-time",
            "title": "Withdrawal Date"
          },
          "reason_category": {
            "type": "string",
            "title": "Reason Category"
          },
          "reason_detail": {
            "type": "string",
            "title": "Reason Detail"
          },
          "scope": {
            "type": "string",
            "title": "Scope"
          },
          "acknowledged_by_investigator": {
            "type": "boolean",
            "title": "Acknowledged By Investigator"
          },
          "investigator_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Investigator Id"
          }
        },
        "type": "object",
        "required": [
          "created_by",
          "reason_for_change",
          "id",
          "study_id",
          "site_id",
          "subject_pseudonym",
          "template_id",
          "withdrawal_date",
          "reason_category",
          "reason_detail",
          "scope",
          "acknowledged_by_investigator",
          "investigator_id"
        ],
        "title": "ConsentWithdrawalResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerifiableCertificateResponse": {
        "properties": {
          "study_id": {
            "type": "string",
            "title": "Study Id"
          },
          "subject_pseudonym": {
            "type": "string",
            "title": "Subject Pseudonym"
          },
          "template_id": {
            "type": "string",
            "title": "Template Id"
          },
          "version_index": {
            "type": "integer",
            "title": "Version Index"
          },
          "html_content": {
            "type": "string",
            "title": "Html Content"
          },
          "digest_sha256": {
            "type": "string",
            "title": "Digest Sha256"
          }
        },
        "type": "object",
        "required": [
          "study_id",
          "subject_pseudonym",
          "template_id",
          "version_index",
          "html_content",
          "digest_sha256"
        ],
        "title": "VerifiableCertificateResponse"
      }
    }
  }
}